Midterm Milestone for Project.
This code can read the touchpad, map the coordinates of the touchpad to
x and y coordinates for the motors. The motors are controlled using a
positional PID that occurs at 200Hz in a timer interrupt. The port D
interrupt vector is called whenever a pin on port D changes and it
calculates the current position of the encoder.